+++ service/idea.go
@@ -8,7 +8,7 @@ import (
 
 // IdeaService contains read and write operations for ideas
 type IdeaService interface {
-	GetAll(tenantID int) ([]*model.Idea, error)
+	GetAll(tenantID int64) ([]*model.Idea, error)
 }
 
 // PostgresIdeaService contains read and write operations for ideas
@@ -17,7 +17,7 @@ type PostgresIdeaService struct {
 }
 
 // GetAll returns all tenant ideas
-func (svc PostgresIdeaService) GetAll(tenantID int) ([]*model.Idea, error) {
+func (svc PostgresIdeaService) GetAll(tenantID int64) ([]*model.Idea, error) {
 	rows, err := svc.DB.Query("SELECT id, title, description FROM ideas WHERE tenant_id = $1", tenantID)
 	if err != nil {
 		return nil, err
