+++ service/tenant_test.go
@@ -31,13 +31,13 @@ func TestTenantService_GetByDomain_Subdomain(t *testing.T) {
 	db, mock, _ := sqlmock.New()
 	defer db.Close()
 
-	rows := sqlmock.NewRows([]string{"id", "name", "subdomain"}).AddRow(234, "My Domain Inc.", "mydomain")
+	rows := sqlmock.NewRows([]string{"id", "name", "subdomain"}).AddRow(int64(234), "My Domain Inc.", "mydomain")
 	mock.ExpectQuery("SELECT id, name, subdomain FROM tenants WHERE subdomain = \\$1 OR cname = \\$2").WithArgs("mydomain", "mydomain").WillReturnRows(rows)
 
 	svc := &service.PostgresTenantService{DB: db}
 	tenant, err := svc.GetByDomain("mydomain")
 
-	Expect(tenant.ID).To(Equal(234))
+	Expect(tenant.ID).To(Equal(int64(234)))
 	Expect(tenant.Name).To(Equal("My Domain Inc."))
 	Expect(tenant.Domain).To(Equal("mydomain.test.canhearyou.com"))
 	Expect(err).To(BeNil())
@@ -49,13 +49,13 @@ func TestTenantService_GetByDomain_FullDomain(t *testing.T) {
 	db, mock, _ := sqlmock.New()
 	defer db.Close()
 
-	rows := sqlmock.NewRows([]string{"id", "name", "subdomain"}).AddRow(234, "My Domain Inc.", "mydomain")
+	rows := sqlmock.NewRows([]string{"id", "name", "subdomain"}).AddRow(int64(234), "My Domain Inc.", "mydomain")
 	mock.ExpectQuery("SELECT id, name, subdomain FROM tenants WHERE subdomain = \\$1 OR cname = \\$2").WithArgs("mydomain", "mydomain.anydomain.com").WillReturnRows(rows)
 
 	svc := &service.PostgresTenantService{DB: db}
 	tenant, err := svc.GetByDomain("mydomain.anydomain.com")
 
-	Expect(tenant.ID).To(Equal(234))
+	Expect(tenant.ID).To(Equal(int64(234)))
 	Expect(tenant.Name).To(Equal("My Domain Inc."))
 	Expect(tenant.Domain).To(Equal("mydomain.test.canhearyou.com"))
 	Expect(err).To(BeNil())
